features/shard: Convert shard block indices to uint64
This patch fixes a crash in FOPs that operate on really large sharded
files where number of participant shards could sometimes exceed
signed int32 max.

The patch also adds GF_ASSERTs to ensure that number of participating
shards is always greater than 0 for files that do have more than one
shard.

Fixes: #1348
(cherry picked from commit cdf01cc47eb2efb427b5855732d9607eec2abc8a)

+++ xlators/features/shard/src/shard.h
@@ -254,9 +254,9 @@ typedef int32_t (*shard_post_update_size_fop_handler_t)(call_frame_t *frame,
 typedef struct shard_local {
     int op_ret;
     int op_errno;
-    int first_block;
-    int last_block;
-    int num_blocks;
+    uint64_t first_block;
+    uint64_t last_block;
+    uint64_t num_blocks;
     int call_count;
     int eexist_count;
     int create_count;
